Multiplayer Player Drawing, bat
Added Player drawing on moving + sending the pos to the server
Added bat to start the server in an endless loop
Added imageFactory that handles every image by itself, and initializes every image aswell
Added logfile + errorlog and two .bat to start the server -> startserver _simple is just starting the server, the other one startserver.bat is adding logs.

Changed index html -> no singleplayer for now

Known Bugs:
Players can "overdraw" other players. When that happens the other player will not be redrawn since his layer is not marked dirty, we need to add a collision detection for that.

Some parts (or all of the map) get redrawn when a player joins. If a player creates a game lobby and then another player joins, the first player will be overdrawn and thus almost disappearing from sight, only leaving the head visible.