Draw fix, Infobar fix, Socket connections
Fixed map not drawn correctly at the beginning of the game. Images have not been loaded when we started the game. Now Images load when opening the page, additionally a callback function is called when the images are done loading.

Added more socket checks. If a player is connected he can create or join a game. Later on i will ad a notification at the popup that notifies the player that no server is available.

Fixed infobar, the infobar was acting weird, turns out the time to fadein and fadeout does not match with the time given to fade out the infobar completely. Fixed.